feat: Adds message input component & removes debug chat in favor of console
