+++ ui/src/chat.module.css
@@ -9,7 +9,6 @@
 .chatBody {
   overflow-y: auto;
   outline: 1px solid #484848;
-  /* outline-offset: 0.25rem; */
   padding: 1rem;
   display: flex;
   flex-direction: column;
@@ -34,30 +33,4 @@
 
 .status.disconnected {
   @apply bg-red-600 text-base;
-}
-
-.messageBox {
-  display: flex;
-  padding: 8px;
-}
-
-.messageInput {
-  flex: 1;
-  padding: 8px;
-  font-size: 16px;
-  border-radius: 4px;
-  border: none;
-}
-
-.sendButton {
-  margin-left: 8px;
-  padding: 8px 16px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  @apply bg-accent text-accent-foreground;
-}
-
-.sendButton:hover {
-  background-color: #0056b3;
 }
