+++ ui/src/components/MessageInput/MessageInput.module.css
@@ -1,0 +1,23 @@
+.messageBox {
+  display: flex;
+  padding: 8px;
+}
+
+.messageInput {
+  flex: 1;
+  padding: 8px;
+  font-size: 16px;
+  border-radius: 4px;
+  border: none;
+  @apply text-foreground;
+}
+
+.sendButton {
+  margin-left: 8px;
+  padding: 8px 16px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  @apply bg-theme-blue text-accent-foreground font-semibold;
+  @apply hover:bg-theme-text hover:text-background;
+}
